DragonFly Movement done
It moves in a zig zag movement. fFlyingSpeed decides the speed it moves on to the screen. fZoomSpeed decides the movement speed when it zooms. fZoomCooldown decides how long before next zoom happens, fRangeX decides how long the dragon fly can move max in its X-position and the fRangeY is how long it can max move in ints Y-position in the zoom.

Fix the spawener maybe.