Wasp is implemented
if XProjectile is empty drag the prefab wasp_prijectile.
if xProjectile Origin is empty drag the WaspProjectileOrigin child from the wasp
F Move Speed is how fast wasp move when it enters the sceen and when it moves randomly when it found it position.
F Radius is how far the wasp is allowed to move before to turn back and move in another direction.
F Cooldown Before Next Move is the time before it starts to move in another direction
F Stand Still Before Fire decides how long it will stand still before firering.
F Fire Rate Decides how fast it fire.